Moved notify_deletion_to_subscribers from message class into PushUpdatesForMessage and consolidated helper methods. Moved specs for upadte_client_store etc into push_updates_for_messages_spec.rb
