+++ app/models/message.rb
@@ -34,29 +34,4 @@ class Message < ActiveRecord::Base
 
   after_destroy :notify_deletion_to_subscribers
 
-  private
-
-  def notify_deletion_to_subscribers
-    send_update self, serialized_user(User.current_user), 'read',
-      admin_channel - donor_channel - browse_channel, ADMIN_APP, :delete
-  end
-
-  def browse_channel
-    return [] unless order
-    Channel.private_channels_for(order.created_by_id, BROWSE_APP)
-  end
-
-  def admin_channel
-    Channel.private_channels_for(User.staff, ADMIN_APP)
-  end
-
-  def donor_channel
-    return [] unless offer
-    Channel.private_channels_for(offer.created_by_id, DONOR_APP)
-  end
-
-  def serialized_user(user)
-    Api::V1::UserSerializer.new(user)
-  end
-
 end
